Herrine's Project/SFY/pjt06/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
Django==3.2.12
django-appconf==1.0.5
django-extensions==3.1.5
django-imagekit==4.1.0
executing==0.8.3
ipython==8.1.1
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.1.0
prompt-toolkit==3.0.28
pure-eval==0.2.2
Pygments==2.11.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
tzdata==2021.5
wcwidth==0.2.5
